tremolo: Fix compiler flags
The purpose of the LOCAL_CFLAGS += -O2 in this file is to make
sure tremolo is built with optimizations enabled - however, if
-O3 is given in the global compiler flags, adding -O2 actually
reduces the optimization level.
Fix this by only adding -O2 if no optimization level is given
globally.

Signed-off-by: Bernhard Rosenkraenzer <Bernhard.Rosenkranzer@linaro.org>

Tremolo: Add workaround for gcc-linaro bug 913528/gcc bug 51794

This adds the -fno-modulo-sched compiler flag (if supported)
to prevent gcc from segfaulting.

This should be reverted once the bug is fixed properly in gcc.

Signed-off-by: Bernhard Rosenkraenzer <Bernhard.Rosenkranzer@linaro.org>